+++ packages/zowe-explorer/__tests__/__unit__/webviews/configEditor/configEditor.unit.test.ts
@@ -2075,6 +2075,7 @@ describe("configEditor", () => {
                 parseErrors: [],
                 secureValuesAllowed: true,
                 tutorialSeen: false,
+                isNewConfig: true,
             });
         });